porting MonitoringExporter test changes from main to 3.4
