+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/ItIstioMonitoringExporter.java
@@ -16,13 +16,10 @@ import oracle.weblogic.kubernetes.annotations.IntegrationTest;
 import oracle.weblogic.kubernetes.annotations.Namespaces;
 import oracle.weblogic.kubernetes.logging.LoggingFacade;
 import oracle.weblogic.kubernetes.utils.ExecResult;
-import org.awaitility.core.ConditionFactory;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import static java.util.concurrent.TimeUnit.MINUTES;
-import static java.util.concurrent.TimeUnit.SECONDS;
 import static oracle.weblogic.kubernetes.TestConstants.ADMIN_PASSWORD_DEFAULT;
 import static oracle.weblogic.kubernetes.TestConstants.ADMIN_USERNAME_DEFAULT;
 import static oracle.weblogic.kubernetes.TestConstants.K8S_NODEPORT_HOST;
@@ -58,7 +55,6 @@ import static oracle.weblogic.kubernetes.utils.OperatorUtils.installAndVerifyOpe
 import static oracle.weblogic.kubernetes.utils.PodUtils.checkPodReady;
 import static oracle.weblogic.kubernetes.utils.SecretUtils.createSecretWithUsernamePassword;
 import static oracle.weblogic.kubernetes.utils.ThreadSafeLogger.getLogger;
-import static org.awaitility.Awaitility.with;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
@@ -81,8 +77,6 @@ class ItIstioMonitoringExporter {
   private static int prometheusPort;
 
   private boolean isPrometheusDeployed = false;
-  // create standard, reusable retry/backoff policy
-  private static ConditionFactory withStandardRetryPolicy = null;
   private static LoggingFacade logger = null;
   private static String oldRegex;
   private static String sessionAppPrometheusSearchKey =
@@ -95,10 +89,6 @@ class ItIstioMonitoringExporter {
   @BeforeAll
   public static void initAll(@Namespaces(3) List<String> namespaces) {
     logger = getLogger();
-    // create standard, reusable retry/backoff policy
-    withStandardRetryPolicy = with().pollDelay(2, SECONDS)
-        .and().with().pollInterval(10, SECONDS)
-        .atMost(6, MINUTES).await();
 
     // get a new unique opNamespace
     logger.info("Assign unique namespace for Operator");
@@ -114,7 +104,7 @@ class ItIstioMonitoringExporter {
     domain2Namespace = namespaces.get(2);
 
     // Label the domain/operator namespace with istio-injection=enabled
-    Map<String, String> labelMap = new HashMap();
+    Map<String, String> labelMap = new HashMap<>();
     labelMap.put("istio-injection", "enabled");
     assertDoesNotThrow(() -> addLabelsToNamespace(domain1Namespace,labelMap));
     assertDoesNotThrow(() -> addLabelsToNamespace(domain2Namespace,labelMap));
@@ -168,7 +158,7 @@ class ItIstioMonitoringExporter {
     // create image with model files
     logger.info("Create image with model file and verify");
 
-    List<String> appList = new ArrayList();
+    List<String> appList = new ArrayList<>();
     appList.add("sessmigr-app");
 
     // build the model file list
@@ -204,7 +194,8 @@ class ItIstioMonitoringExporter {
           "Can't modify Prometheus CM, not possible to monitor " + domainUid);
     }
     //verify metrics via prometheus
-    checkMetricsViaPrometheus(searchKey, "sessmigr", prometheusPort);
+    checkMetricsViaPrometheus(searchKey, "sessmigr",
+        K8S_NODEPORT_HOST + ":" + prometheusPort);
   }
 
   /**
@@ -214,7 +205,7 @@ class ItIstioMonitoringExporter {
     // create image with model files
     logger.info("Create image with model file with monitoring exporter app and verify");
 
-    List<String> appList = new ArrayList();
+    List<String> appList = new ArrayList<>();
     appList.add(monexpAppDir);
     appList.add("sessmigr-app");
 
@@ -264,7 +255,7 @@ class ItIstioMonitoringExporter {
         String.format("createSecret failed for %s", encryptionSecretName));
 
     // create WDT config map without any files
-    createConfigMapAndVerify(configMapName, domainUid, domainNamespace, Collections.EMPTY_LIST);
+    createConfigMapAndVerify(configMapName, domainUid, domainNamespace, Collections.emptyList());
 
     // create the domain object
     Domain domain = createIstioDomainResource(domainUid,
@@ -307,7 +298,7 @@ class ItIstioMonitoringExporter {
 
     String clusterService = domainUid + "-cluster-" + clusterName + "." + domainNamespace + ".svc.cluster.local";
 
-    Map<String, String> templateMap  = new HashMap();
+    Map<String, String> templateMap  = new HashMap<>();
     templateMap.put("NAMESPACE", domainNamespace);
     templateMap.put("DUID", domainUid);
     templateMap.put("ADMIN_SERVICE",adminServerPodName);
