+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/actions/impl/PrometheusParams.java
@@ -47,6 +47,10 @@ public class PrometheusParams {
     return helmParams;
   }
 
+  public int getNodePortServer() {
+    return nodePortServer;
+  }
+
   /**
    * Loads Helm values into a value map.
    *
