+++ integration-tests/src/test/java/oracle/weblogic/kubernetes/assertions/impl/Prometheus.java
@@ -12,16 +12,17 @@ public class Prometheus {
   /**
    * Check if the prometheus pods are running in a given namespace.
    * @param namespace in which to check for the prometheus pods
+   * @param releaseName prometheus release name
    * @return true if found and running otherwise false
    */
-  public static Callable<Boolean> isReady(String namespace) {
+  public static Callable<Boolean> isReady(String namespace, String releaseName) {
     Map<String,String> labelMapPromSvc = new HashMap<>();
     labelMapPromSvc.put("component", "server");
     Map<String,String> labelMapAlertMgr = new HashMap<>();
     labelMapAlertMgr.put("component", "alertmanager");
     return () -> {
-      return (Kubernetes.isPodReady(namespace, labelMapAlertMgr, "prometheus-alertmanager")
-              && Kubernetes.isPodReady(namespace, labelMapPromSvc, "prometheus-server"));
+      return (Kubernetes.isPodReady(namespace, labelMapAlertMgr, releaseName + "-alertmanager")
+              && Kubernetes.isPodReady(namespace, labelMapPromSvc, releaseName + "-server"));
     };
   }
 }
